Linux packaging fix: use "data-plotter" final executable name instead of "DataPlotter"
The executable name chosen for the GNU/Linux build was "data-plotter"
even before my first attempt for proper Linux packaging so I have
tried to keep it. The option to use "DataPlotter" name even
on Linux (same as on Windows) is possible, but CammelCase it is
much less common in the /usr/bin. In the fact I have no such binary
on my system.

Automatic lover case to "dataplotter" is possible but I think
that it is much less distinguishable.

Corrected missing executable install as well
  install(TARGETS DataPlotter)
to install final executable.

The SPEC file for Fedora and SUSE cleaned according to Lubos Kocman
review of QtRvSim SPEC at  InstallFest 2025.

Signed-off-by: Pavel Pisa <pisa@fel.cvut.cz>